+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `LiftA2` to lift and apply a `Bifunction` to two `Applicative`s
 - `Flatten` to lazily flatten nested `Iterable<Iterable<A>>`s to `Iterable<A>`
 - `Replicate`, short-hand composition of `take` and `repeat`
+- `Distinct` to produce an `Iterable` of distinct values in another `Iterable` 
 
 ## [1.6.2] - 2017-08-20
 ### Fixed